tools/geneos: start work on rebuilding `ps`/`status` command as the code is a maze and hard to enhance
pkg/process: lots of refactoring and rewriting as a result of the above changes

+++ pkg/process/files.go
@@ -3,7 +3,7 @@ package process
 import "io/fs"
 
 type ProcessFDs struct {
-	PID   int
+	PID   int64
 	FD    int
 	Path  string
 	Lstat fs.FileInfo
