+++ pkg/process/files_windows.go
@@ -2,6 +2,6 @@ package process
 
 import "github.com/itrs-group/cordial/pkg/host"
 
-func OpenFiles(h host.Host, pid int) (files []ProcessFDs) {
+func OpenFiles(h host.Host, pid int64) (files []ProcessFDs) {
 	return
 }
